cloud_functions/requirements.txt
google-cloud-storage==2.11.0
google-cloud-bigquery==3.11.2
xmltodict==0.13.0
plotly